Add distance model training and eval
Prompt:
I want to train a learned representation for image translation.  The representation should be useful for pixel distance (or semantic distance, like an action distance in a reinforcement learning context.)

As part of the learning process, use appropriate data augmentations, like crop, scale, add noise, etc.

Create a script that takes the input and trains a PyTorch model.  The model should take in frames of size 240x224.

There should be two scripts: a training script and an evaluation script.

=== For the training script ===

The input is a folder of trajectories, organized like:
traj_dumps/traj_<n>/states/state_<n>.png

The input images are of size 240x224.

Use PyTorch to train a model that outputs the image distance between two frames.  Use augmentations of the same image to create data with groundtruth.

For example, take image from state_0 and then state_0 and add noise over the whole image.  Treat those as labeled pairs with distance 0.
For another example, take image from state_0, then crop and shift the image (and add padding noise to ensure the resulting image stays at 240x224).  The state_0 and state_0_cropped would then be pairs with a distance of the shift amount.  Use combinations of noise, cropping, shifting, and some scaling up to about 20% of the image size.

Output a sampling of the training pairs, but mostly do the augmentations in memory and don’t store all the results.

As the model trains, there should be periodic outputs of the model accuracy/loss.

=== For the evaluation script ===

Given the same input trajectories, the model should be used to produce outputs:
* A list of distance errors between the groundtruth augmented distance and the image distance.
* A histogram of the image distance errors
* A list of distance errors between the inferred distance and the distance between steps of the trajectory.  For example, state_3 and state_6 will have an action distance of “3”.
* A histogram of the error between steps of the trajectory and image distance.