Image distance is now a vector, not scalar
Prompt:
When outputting debug image pairs, also run inference on each pair to show how the model estimated the distance.  Run inference two ways: from A->B and B->A.  In the A picture, draw a line from the center of the image A to where the model thinks the image A moved to in B.  In the B picture, draw a line from the center of the image B to where the model thinks the image B moved to to in A.

Prompt:
The model output should already be a motion vector.  How is the output distance different than a motion vector?  I thought we were visualizing the output of the current model, not doing a search over distances for a vector?