Use original and augmented images as both A->B and B->A pairs for training
Prompt:
In the debug output, I'm seeing only the 2nd image apply a crop/translation/padding.  But, I thought the crop/translation/padding should be randomly applied to both the A image and the B image.  Either that, or the pairs should be used as bidirectional training examples, from A->B and B->A.

Prompt:
Check the canvas again, make sure it's correct