Add script for evaluating image distance vs action steps
Prompt:
Write another eval script that compares the inferred pixel distance to action distance. The model should accept a pair of images from a trajectory (or occasionally different trajectories). The output should be pixel distance, as originally trained. But, the recorded output error should be the inferred pixel distance, and the groundtruth should be the number of trajectory steps between the images. If the images are from different trajectories, then the number of steps should be some kind of "unknown" output. Then, after the list of eval tuples (inferred=pixel distance, groundtruth=trajectory steps) is available, make a scatter plot of trajectory steps (on x axis) vs pixel distance (on y axis).

For unknown/blank/Nan from correlation stats, include it as a separate stat.  And, create a histogram of the image distance for the unknown step gap cases.