Add action distance training and eval scripts
Prompt:
I want to train a learned representation for image translation.  The representation should be useful for pixel distance (or semantic distance, like an action distance in a reinforcement learning context.)

As part of the learning process, use data augmentations that add noise (different types, and potentially a lot of noise) and gaussian blur at varying amounts and different focus points in the image.

Create a script that takes the input and trains a PyTorch model.  The model should take in frames of size 240x224.

There should be two scripts: a training script and an evaluation script.

=== For the training script ===

The input is a folder of trajectories, organized like:
traj_dumps/traj_<n>/states/state_<n>.png

The input images are of size 240x224.

Use PyTorch to train a model that outputs the action distance between two frames.  Use augmentations of the same image to create data with groundtruth.

For example, take image from state_0 and then state_0 and add noise over the each image.  Treat those as labeled pairs with distance based on their distance in the trajectory.  For the same state with different noise amounts, that would be an A and B with distance 0.  For different states in a trajectory, say state_3 and state_6, each would add noise and then they would have groundtruth action distance of 3.

Output a sampling of the training pairs (debug image pairs), but mostly do the augmentations in memory and don’t store all the results.

When outputting debug image pairs, also run inference on each pair to show how the model estimated the latent action distance.  Run inference two ways: from A->B and B->A.  In the A picture, draw a line from the center of the image A to where the model thinks the image A moved to in B.  In the B picture, draw a line from the center of the image B to where the model thinks the image B moved to to in A.

As the model trains, there should be periodic outputs of the model accuracy/loss.

The action distance output of the model should be a vector.  Very loosely the action distance magnitude should be the number of steps between states.  The direction of the vector should be the direction to move in the latent space to get from (say) state_3 to state_6.

=== For the evaluation script ===

Given the same input trajectories, the model should be used to produce outputs:
* A list of distance errors between the groundtruth augmented distance and the image distance.
* A histogram of the image distance errors
* A list of distance errors between the inferred distance and the distance between steps of the trajectory.  For example, state_3 and state_6 will have an action distance of “3”.

=== Plan ===

Before writing code, give me a plan with an outline of the scripts, and choice of algorithms and models.  Highlight anything that is unclear and also highlight any import assumptions that are being made.
* A histogram of the error between steps of the trajectory and image distance.

Prompt:
Update the plan with the following changes:
* Call this set of training and eval "action_distance_train.py" and "action_distance_eval.py"
* Avoid directly training for pixel distance.  There should be no crop/translation/etc. used for augmentation.  We want the model to learn important features for moving in action distance only.  The pixel difference may not matter.
* Think about if setting this problem up as a flow matching model would better learn any of the parameters.
* For training, we want to check that the model infers correct action distance between states, not pixel distance.  To test, we'll want to feed in pairs of states within the same trajectory, and get out a minimum possible distance between the states (which may be less than the actual distance in the trajectory).

Show me the new plan before writing any code.