Update debug drawing for action distance
Prompt:
In debug images, put debug text on the bottom of the image not the top. And, draw the vector as a line from the center of the image out in the direction of the latent space according to the vector. Scale the debug visualization so that an output vector of magnitude "1 action step" is equivalent to 10 pixels in the image.