Add visualization output for debugging
Prompt:
The direction of the latent vector doesn't seem to be meaningful when projected into 2d and drawn on the image.  Give me some better ideas for how I can meaningfully visualize the latent vector between two images.

Prompt:
Yes, update the script to give an output visualization for each of the suggested views and put them in out.action_distance/vis/ as described.